# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KindlePool API",
    "version": "0.1.0",
    "description": "REST API for KindlePool — micro-sponsor pools on Stellar Soroban.\n\nBase URL: `https://api.kindlepool.dev/v1`\n\nAuthentication: `X-API-Key` header (get one from the developer portal)."
  },
  "servers": [
    { "url": "https://api.kindlepool.dev", "description": "Production" },
    { "url": "https://staging-api.kindlepool.dev", "description": "Staging" },
    { "url": "http://localhost:3001", "description": "Local development" }
  ],
  "security": [],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": { "type": "apiKey", "in": "header", "name": "X-API-Key" }
    },
    "schemas": {
      "Pool": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "contract_id": { "type": "string" },
          "creator": { "type": "string" },
          "token": { "type": "string" },
          "goal": { "type": "string" },
          "total_deposited": { "type": "string" },
          "deadline": { "type": "integer" },
          "status": { "type": "string", "enum": ["open", "awaiting_vote", "paid", "expired", "disputed", "appealed"] },
          "work_hash": { "type": "string", "nullable": true },
          "vote_deadline": { "type": "integer", "nullable": true },
          "yes_votes": { "type": "string" },
          "no_votes": { "type": "string" },
          "metadata_hash": { "type": "string" },
          "total_supporters": { "type": "integer" },
          "created_at": { "type": "integer" },
          "updated_at": { "type": "integer" }
        }
      },
      "Supporter": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "pool_id": { "type": "integer" },
          "address": { "type": "string" },
          "amount": { "type": "string" },
          "voted": { "type": "boolean" },
          "created_at": { "type": "integer" }
        }
      },
      "Event": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "pool_id": { "type": "integer" },
          "event_type": { "type": "string" },
          "data": { "type": "string" },
          "ledger": { "type": "integer" },
          "ts": { "type": "integer" }
        }
      },
      "PaginatedPools": {
        "type": "object",
        "properties": {
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/Pool" } },
          "page": { "type": "integer" },
          "limit": { "type": "integer" },
          "total": { "type": "integer" },
          "total_pages": { "type": "integer" }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      }
    }
  },
  "paths": {
    "/api/v1/health": {
      "get": {
        "summary": "Health check",
        "tags": ["System"],
        "security": [],
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "object", "properties": { "status": { "type": "string" }, "timestamp": { "type": "integer" } } } } } }
        }
      }
    },
    "/api/v1/pools": {
      "get": {
        "summary": "List pools",
        "tags": ["Pools"],
        "security": [],
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["open", "awaiting_vote", "paid", "expired"] } },
          { "name": "creator", "in": "query", "schema": { "type": "string" } },
          { "name": "sort", "in": "query", "schema": { "type": "string", "enum": ["newest", "ending_soon", "most_funded"] } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20, "maximum": 100 } }
        ],
        "responses": {
          "200": { "description": "Paginated pool list", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PaginatedPools" } } } }
        }
      }
    },
    "/api/v1/pools/{id}": {
      "get": {
        "summary": "Get pool by ID",
        "tags": ["Pools"],
        "security": [],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": { "description": "Pool details", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Pool" } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/v1/pools/{id}/supporters": {
      "get": {
        "summary": "List pool supporters",
        "tags": ["Pools"],
        "security": [],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": { "description": "List of supporters", "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/Supporter" } } } } } } }
        }
      }
    },
    "/api/v1/pools/{id}/events": {
      "get": {
        "summary": "Get pool event history",
        "tags": ["Pools"],
        "security": [],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50 } }
        ],
        "responses": {
          "200": { "description": "List of events", "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/Event" } } } } } } }
        }
      }
    },
    "/api/v1/supporters/{address}/pools": {
      "get": {
        "summary": "Get pools funded by a supporter",
        "tags": ["Supporters"],
        "security": [],
        "parameters": [{ "name": "address", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "List of pools" }
        }
      }
    },
    "/api/v1/creators/{address}/pools": {
      "get": {
        "summary": "Get pools created by a creator",
        "tags": ["Creators"],
        "security": [],
        "parameters": [{ "name": "address", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "List of pools" }
        }
      }
    },
    "/api/v1/events": {
      "get": {
        "summary": "Get all events",
        "tags": ["Events"],
        "security": [],
        "parameters": [
          { "name": "type", "in": "query", "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50 } }
        ],
        "responses": {
          "200": { "description": "List of events" }
        }
      }
    }
  }
}